requirements.txt
nonebot2==2.0.0b2
nonebot-adapter-onebot==2.0.0b1
nb-cli==0.6.6
opencv-python
numpy
basicsr>=1.3.3.11
torch>=1.7
torchvision
tqdm